Pager fixes

+++ src/main/java/gr/dsigned/springcrudutils/ParameterizedCRUDController.java
@@ -32,7 +32,7 @@ public class ParameterizedCRUDController<E extends SystemEntity> {
         int pageNo = page == null ? 0 : (page.intValue());
         int offset = pageNo * sizeNo;
         Long itemNo = dao.countAll(entityClass);
-        Pager pager = new Pager(request.getRequestURI(), itemNo.longValue(), Integer.valueOf(pageNo).longValue(), sizeNo);
+        Pager pager = new Pager(request.getRequestURI(), itemNo.intValue(), pageNo, sizeNo);
         model.addAttribute("itemNo", itemNo);
         model.addAttribute("sizeNo", sizeNo);
         model.addAttribute("pageNo", pageNo);
